fix: resolve Postgres-specific test-server failures

+++ server/repositories/network.repo.ts
@@ -32,7 +32,7 @@ export class NetworkRepo {
           isCentral: s.derivedNodes.isCentral,
           sentiment: s.derivedNodes.sentiment,
           degree:
-            sql<number>`COALESCE((SELECT COUNT(*) FROM node_links nl WHERE nl.source_slug = derived_nodes.slug OR nl.target_slug = derived_nodes.slug), 0)`.as(
+            sql<number>`COALESCE((SELECT COUNT(*)::int FROM node_links nl WHERE nl.source_slug = derived_nodes.slug OR nl.target_slug = derived_nodes.slug), 0)`.as(
               'degree',
             ),
         })
